modules/search.py
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import (TimeoutException, 
                                      NoSuchElementException, 
                                      WebDriverException)
import time
import yaml
import os
from enum import Enum, auto
from typing import Dict, List, Optional, Union
from dataclasses import dataclass

class SearchType(Enum):
    """Enum representing different LinkedIn search types"""
    PEOPLE = auto()
    JOBS = auto()
    COMPANIES = auto()
    CONTENT = auto()

@dataclass
class SearchResult:
    """Data class for storing search results"""
    title: str
    subtitle: str
    link: str
    metadata: Optional[Dict] = None

class LinkedInAutomation:
    """
    Comprehensive LinkedIn automation tool with:
    - Login functionality
    - Multiple search types
    - Advanced filtering
    - Robust error handling
    """
    
    DEFAULT_CONFIG = {
        'login_url': 'https://www.linkedin.com/login',
        'search_urls': {
            'people': 'https://www.linkedin.com/search/results/people/',
            'jobs': 'https://www.linkedin.com/jobs/search/',
            'companies': 'https://www.linkedin.com/search/results/companies/',
            'content': 'https://www.linkedin.com/search/results/content/'
        },
        'limits': {
            'max_results': 100,
            'delay_between_actions': 2
        }
    }

    def __init__(self, browser_manager, config_path: str = 'config/config.yaml'):
        """
        Initialize LinkedIn automation tool
        
        Args:
            browser_manager: Browser management instance
            config_path: Path to configuration file
        """
        self.browser = browser_manager
        self.config = self._load_config(config_path)
        self._setup_locators()
        self._logged_in = False

    def _load_config(self, config_path: str) -> Dict:
        """Load and merge configuration from file with defaults"""
        try:
            with open(config_path, 'r') as f:
                user_config = yaml.safe_load(f) or {}
            config = {**self.DEFAULT_CONFIG, **(user_config.get('linkedin', {}))}
            self._validate_config(config)
            return config
        except Exception as e:
            raise Exception(f"Configuration error: {str(e)}")

    def _validate_config(self, config: Dict) -> None:
        """Validate required configuration values"""
        if not config.get('credentials', {}).get('username'):
            raise ValueError("Username not configured")
        if not config.get('credentials', {}).get('password'):
            raise ValueError("Password not configured")

    def _setup_locators(self) -> None:
        """Initialize all element locators with fallbacks"""
        self.locators = {
            'login': {
                'username': [(By.ID, 'username'), (By.NAME, 'session_key')],
                'password': [(By.ID, 'password'), (By.NAME, 'session_password')],
                'submit': [(By.XPATH, "//button[@type='submit']")]
            },
            'search': {
                'input': {
                    'people': [(By.XPATH, "//input[contains(@aria-label, 'Search by name')]")],
                    'jobs': [(By.XPATH, "//input[contains(@aria-label, 'Search job titles')]")],
                    'companies': [(By.XPATH, "//input[contains(@aria-label, 'Search companies')]")],
                    'content': [(By.XPATH, "//input[contains(@aria-label, 'Search content')]")]
                },
                'results': {
                    'people': (By.XPATH, "//li[contains(@class, 'reusable-search__result-container')]"),
                    'jobs': (By.XPATH, "//li[contains(@class, 'jobs-search-results__list-item')]"),
                    'companies': (By.XPATH, "//li[contains(@class, 'reusable-search__result-container')]"),
                    'content': (By.XPATH, "//div[contains(@class, 'search-content__result-container')]")
                }
            },
            'navigation': {
                'next_page': (By.XPATH, "//button[contains(@aria-label, 'Next')]")
            }
        }

    def perform_search(self, 
                      search_type: Union[SearchType, str],
                      query: Optional[str] = None,
                      filters: Optional[Dict] = None,
                      max_results: Optional[int] = None) -> List[SearchResult]:
        """
        Perform LinkedIn search with optional filters
        
        Args:
            search_type: Type of search (PEOPLE, JOBS, COMPANIES, CONTENT)
            query: Search query string
            filters: Dictionary of filters to apply
            max_results: Maximum number of results to return
            
        Returns:
            List of SearchResult objects
        """
        if isinstance(search_type, str):
            search_type = SearchType[search_type.upper()]
            
        if not self._logged_in:
            self.login()

        self._navigate_to_search_page(search_type)
        
        if query:
            self._enter_search_query(search_type, query)
            
        if filters:
            self._apply_filters(search_type, filters)
            
        return self._collect_results(search_type, max_results)

    def login(self) -> None:
        """Login to LinkedIn with configured credentials"""
        try:
            self.browser.driver.get(self.config['login_url'])
            self._wait_for_page_load()

            # Find working locators for login fields
            username_locator = self._find_working_locator(self.locators['login']['username'])
            password_locator = self._find_working_locator(self.locators['login']['password'])
            submit_locator = self._find_working_locator(self.locators['login']['submit'])

            creds = self.config['credentials']
            self._safe_send_keys(username_locator, creds['username'])
            self._safe_send_keys(password_locator, creds['password'])
            self._safe_click(submit_locator)

            if not self._verify_login():
                raise Exception("Login verification failed")
                
            self._logged_in = True
            
        except Exception as e:
            self._capture_debug_info("login_failed")
            raise Exception(f"Login failed: {str(e)}")

    def _navigate_to_search_page(self, search_type: SearchType) -> None:
        """Navigate to appropriate search page"""
        page_types = {
            SearchType.PEOPLE: 'people',
            SearchType.JOBS: 'jobs',
            SearchType.COMPANIES: 'companies',
            SearchType.CONTENT: 'content'
        }
        url = self.config['search_urls'][page_types[search_type]]
        self.browser.driver.get(url)
        self._wait_for_page_load()

    def _enter_search_query(self, search_type: SearchType, query: str) -> None:
        """Enter search query into appropriate search field"""
        input_locator = self._find_working_locator(
            self.locators['search']['input'][search_type.name.lower()])
        
        self._safe_clear(input_locator)
        self._safe_send_keys(input_locator, query)
        self._safe_send_keys(input_locator, Keys.RETURN)
        self._wait_for_page_load()

    def _apply_filters(self, search_type: SearchType, filters: Dict) -> None:
        """Apply search filters based on search type"""
        # Implementation would vary by search type
        pass

    def _collect_results(self, 
                        search_type: SearchType,
                        max_results: Optional[int] = None) -> List[SearchResult]:
        """Collect and return search results"""
        max_results = max_results or self.config['limits']['max_results']
        results = []
        results_locator = self.locators['search']['results'][search_type.name.lower()]
        
        while len(results) < max_results:
            # Scroll to load more results
            self._scroll_to_bottom()
            
            # Find all result elements
            result_elements = self._safe_find_elements(results_locator)
            
            # Process new results
            for element in result_elements[len(results):max_results]:
                try:
                    result = self._extract_result(element, search_type)
                    if result:
                        results.append(result)
                except Exception as e:
                    print(f"Error processing result: {str(e)}")
                    continue
            
            # Check for more results
            if len(results) >= max_results or not self._has_more_results():
                break
                
            # Go to next page if available
            self._go_to_next_page()
            
        return results[:max_results]

    def _extract_result(self, element, search_type: SearchType) -> Optional[SearchResult]:
        """Extract information from a single result element"""
        extractors = {
            SearchType.PEOPLE: self._extract_person_result,
            SearchType.JOBS: self._extract_job_result,
            SearchType.COMPANIES: self._extract_company_result,
            SearchType.CONTENT: self._extract_content_result
        }
        return extractors[search_type](element)

    # Helper methods would follow...
    # _find_working_locator, _safe_send_keys, _wait_for_page_load, etc.

if __name__ == "__main__":
    # Example usage
    from browser_manager import BrowserManager
    
    browser = BrowserManager()
    linkedin = LinkedInAutomation(browser)
    
    try:
        # Search for people
        people = linkedin.perform_search(
            search_type="PEOPLE",
            query="Software Engineer",
            filters={"location": "San Francisco"},
            max_results=50
        )
        
        # Search for jobs
        jobs = linkedin.perform_search(
            search_type=SearchType.JOBS,
            query="Python Developer",
            filters={"remote": True, "experience": "mid-senior"}
        )
        
    finally:
        browser.quit()