requirements.txt
selenium==4.1.0
pytesseract==0.3.9
pillow==9.0.1
python-dotenv==0.19.2
opencv-python-headless==4.5.5.64
numpy==1.22.3
PyYAML==6.0.1
webdriver_manager>=4.0.0
